requirements.txt
torch>=2.0.0
mrcfile>=1.4.0
Pillow>=9.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
numpy>=1.24.0
